change layout structure and linked buat-akun-cabang form
