+++ app/account-block/page.jsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 
 const AccountBlock = () => {
   return (
-    <>
+    <div className="flex flex-col gap-[20px]">
       <h2 className="text-2xl font-bold text-black">Account Block</h2>
       <div className="flex h-[48px] gap-[18px]">
         <div className="w-3/5 relative">
@@ -50,7 +50,7 @@ const AccountBlock = () => {
               fill="white"
             />
           </svg>
-          <p className="ml-5 text-base font-regular">Tambah Cabang</p>
+          <p className="ml-5 text-base font-regular">Tambah Akun</p>
         </button>
         <button className="w-1/5 bg-primary px-[15px] py-[8px] flex justify-center items-center text-white rounded-md">
           <svg
@@ -191,7 +191,7 @@ const AccountBlock = () => {
           </tbody>
         </table>
       </div>
-    </>
+    </div>
   );
 };
