--- app/buat-akun-cabang/page.jsx
@@ -1,559 +1,0 @@
-import React from "react";
-import Image from "next/image";
-
-const BuatAkunCabang = () => {
-  return (
-    <main className="flex h-full bg-mainBg">
-      <div className="w-[280px] flex flex-col justify-between p-[28px] bg-primary fixed h-screen">
-        <div className="flex flex-col">
-          <div className="flex items-center mb-[60px]">
-            <div className="w-[35px] h-[40px] bg-filled-color rounded-md mr-[15px]"></div>
-            <div className="flex flex-col">
-              <h3 className="text-white font-bold text-lg">KSP</h3>
-              <p className="text-white font-regular text-base">
-                Sentosa Makmur
-              </p>
-            </div>
-          </div>
-          <div className="bg-[#d9d9d9] mb-[16px] py-[14px] px-[16px] flex items-center rounded-md hover:bg-secondary">
-            <svg
-              width="20"
-              height="20"
-              viewBox="0 0 20 20"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-              className="mr-[17px]"
-            >
-              <path
-                d="M11.1111 6.66667V0H20V6.66667H11.1111ZM0 11.1111V0H8.88889V11.1111H0ZM11.1111 20V8.88889H20V20H11.1111ZM0 20V13.3333H8.88889V20H0ZM2.22222 8.88889H6.66667V2.22222H2.22222V8.88889ZM13.3333 17.7778H17.7778V11.1111H13.3333V17.7778ZM13.3333 4.44444H17.7778V2.22222H13.3333V4.44444ZM2.22222 17.7778H6.66667V15.5556H2.22222V17.7778Z"
-                fill="black"
-              />
-            </svg>
-            <p className="text-black text-base font-medium">Dashboard</p>
-          </div>
-          <div className="mb-[16px] py-[14px] px-[16px] flex items-center rounded-md hover:bg-secondary">
-            <svg
-              width="20"
-              height="18"
-              viewBox="0 0 20 18"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-              className="mr-[17px]"
-            >
-              <path
-                d="M19.9923 4.76872L18.5648 0.48631C18.5152 0.338581 18.4185 0.211167 18.2896 0.123529C18.1607 0.0358913 18.0067 -0.00715311 17.8511 0.000970646H2.14892C1.99328 -0.00715311 1.83925 0.0358913 1.71037 0.123529C1.58149 0.211167 1.48484 0.338581 1.43518 0.48631L0.00771216 4.76872C-0.00257072 4.8445 -0.00257072 4.92133 0.00771216 4.99711V9.27952C0.00771216 9.46882 0.082909 9.65036 0.21676 9.78421C0.350611 9.91806 0.532153 9.99326 0.721447 9.99326H1.43518V17.1306H2.86265V9.99326H7.14506V17.1306H18.5648V9.99326H19.2786C19.4678 9.99326 19.6494 9.91806 19.7832 9.78421C19.9171 9.65036 19.9923 9.46882 19.9923 9.27952V4.99711C20.0026 4.92133 20.0026 4.8445 19.9923 4.76872ZM17.1373 15.7031H8.57253V9.99326H17.1373V15.7031ZM18.5648 8.56579H15.7099V5.71085H14.2824V8.56579H10.7137V5.71085H9.28627V8.56579H5.71759V5.71085H4.29012V8.56579H1.43518V5.11131L2.66281 1.42844H17.3372L18.5648 5.11131V8.56579Z"
-                fill="#F8FAFC"
-              />
-            </svg>
-            <p className="text-white text-base font-medium">Data Master</p>
-          </div>
-          <div className="mb-[16px] py-[14px] px-[16px] flex items-center rounded-md hover:bg-secondary">
-            <svg
-              width="20"
-              height="20"
-              viewBox="0 0 20 20"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-              className="mr-[17px]"
-            >
-              <path
-                d="M11.2109 11.25C10.6706 10.8464 10.0879 10.5371 9.46289 10.3223C8.83789 10.1074 8.18359 10 7.5 10C6.92708 10 6.3737 10.0749 5.83984 10.2246C5.30599 10.3743 4.80794 10.5827 4.3457 10.8496C3.88346 11.1165 3.46354 11.4421 3.08594 11.8262C2.70833 12.2103 2.38281 12.6335 2.10938 13.0957C1.83594 13.5579 1.62435 14.056 1.47461 14.5898C1.32487 15.1237 1.25 15.6771 1.25 16.25H0C0 15.4688 0.113932 14.7168 0.341797 13.9941C0.569661 13.2715 0.898438 12.6042 1.32812 11.9922C1.75781 11.3802 2.26562 10.8366 2.85156 10.3613C3.4375 9.88607 4.10156 9.51172 4.84375 9.23828C4.47917 9.01042 4.15365 8.74349 3.86719 8.4375C3.58073 8.13151 3.33333 7.79297 3.125 7.42188C2.91667 7.05078 2.76367 6.66341 2.66602 6.25977C2.56836 5.85612 2.51302 5.4362 2.5 5C2.5 4.3099 2.63021 3.66211 2.89062 3.05664C3.15104 2.45117 3.50911 1.92383 3.96484 1.47461C4.42057 1.02539 4.95117 0.667318 5.55664 0.400391C6.16211 0.133464 6.8099 0 7.5 0C8.1901 0 8.83789 0.130208 9.44336 0.390625C10.0488 0.651042 10.5762 1.00911 11.0254 1.46484C11.4746 1.92057 11.8327 2.45117 12.0996 3.05664C12.3665 3.66211 12.5 4.3099 12.5 5C12.5 5.42969 12.4479 5.84961 12.3438 6.25977C12.2396 6.66992 12.0833 7.05729 11.875 7.42188C11.6667 7.78646 11.4193 8.12175 11.1328 8.42773C10.8464 8.73372 10.5176 9.00391 10.1465 9.23828C10.5371 9.38802 10.9049 9.5638 11.25 9.76562V11.25H11.2109ZM3.75 5C3.75 5.52083 3.84766 6.00586 4.04297 6.45508C4.23828 6.9043 4.50521 7.30143 4.84375 7.64648C5.18229 7.99154 5.57943 8.26172 6.03516 8.45703C6.49089 8.65234 6.97917 8.75 7.5 8.75C8.01432 8.75 8.49935 8.65234 8.95508 8.45703C9.41081 8.26172 9.80794 7.99479 10.1465 7.65625C10.485 7.31771 10.7552 6.92057 10.957 6.46484C11.1589 6.00911 11.2565 5.52083 11.25 5C11.25 4.48568 11.1523 4.00065 10.957 3.54492C10.7617 3.08919 10.4948 2.69206 10.1562 2.35352C9.81771 2.01497 9.41732 1.74479 8.95508 1.54297C8.49284 1.34115 8.00781 1.24349 7.5 1.25C6.97917 1.25 6.49414 1.34766 6.04492 1.54297C5.5957 1.73828 5.19857 2.00521 4.85352 2.34375C4.50846 2.68229 4.23828 3.08268 4.04297 3.54492C3.84766 4.00716 3.75 4.49219 3.75 5ZM20 12.5V20H10V12.5H12.5V10H17.5V12.5H20ZM13.75 12.5H16.25V11.25H13.75V12.5ZM18.75 16.25H17.5V17.5H16.25V16.25H13.75V17.5H12.5V16.25H11.25V18.75H18.75V16.25ZM18.75 13.75H11.25V15H18.75V13.75Z"
-                fill="#F8FAFC"
-              />
-            </svg>
-            <p className="text-white text-base font-medium ">Account Block</p>
-          </div>
-        </div>
-        <div className="flex flex-col">
-          <button className="bg-white/10  mb-[16px] py-[14px] px-[16px] flex items-center rounded-md">
-            <svg
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-              className="mr-[12px]"
-            >
-              <path
-                d="M8.3335 16.1673L12.5002 12.0007L8.3335 7.83398"
-                stroke="white"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-              <path
-                d="M12.5 12H2.5"
-                stroke="white"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-              <path
-                d="M12.5 4.5H15.8333C16.2754 4.5 16.6993 4.67559 17.0118 4.98816C17.3244 5.30072 17.5 5.72464 17.5 6.16667V17.8333C17.5 18.2754 17.3244 18.6993 17.0118 19.0118C16.6993 19.3244 16.2754 19.5 15.8333 19.5H12.5"
-                stroke="white"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </svg>
-            <p className="text-mainBg text-base font-bold">Keluar</p>
-          </button>
-          <button className="flex items-center border-t border-white/10 pt-[15px] justify-between">
-            <div className="flex items-center">
-              <Image
-                src={"/images/Image.png"}
-                alt="image"
-                width={32}
-                height={32}
-                className="mr-[12px]"
-              />
-              <div className="flex flex-col items-start">
-                <p className="text-white text-base font-semibold ">Teller</p>
-                <p className="text-white text-sm font-regular">Indah sari</p>
-              </div>
-            </div>
-            <svg
-              width="4"
-              height="20"
-              viewBox="0 0 4 20"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <circle cx="2" cy="2" r="2" fill="white" />
-              <circle cx="2" cy="10" r="2" fill="white" />
-              <circle cx="2" cy="18" r="2" fill="white" />
-            </svg>
-          </button>
-        </div>
-      </div>
-      <form action="">
-        <div className="m-[30px] flex flex-col gap-[10px] ml-[310px]">
-          <h2 className="text-2xl font-bold text-black">Buat Akun Cabang</h2>
-          <div className="bg-white p-[20px] rounded-xl">
-            <p className="text-black font-regular text-lg mb-[10px]">
-              Silahkan lengkapi form berikut dengan benar.
-            </p>
-            <div className="flex flex-col gap-2">
-              <div className="flex-1">
-                <label htmlFor="idCabang">ID Cabang</label>
-                <input
-                  type="text"
-                  id="idCabang"
-                  name="idCabang"
-                  placeholder="Auto generate"
-                  className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] italic bg-white"
-                  disabled
-                />
-              </div>
-              <div className="flex gap-3">
-                <div className="w-2/5">
-                  <div>
-                    <label htmlFor="alamatCabang">Alamat Lengkap Cabang</label>
-                    <input
-                      type="text"
-                      id="alamatCabang"
-                      name="alamatCabang"
-                      placeholder="Isi alamat lengkap cabang"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                </div>
-                <div className="flex w-3/5 gap-3">
-                  <div>
-                    <label htmlFor="kelurahan">Kelurahan/Desa</label>
-                    <input
-                      type="text"
-                      id="kelurahan"
-                      name="kelurahan"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                  <div>
-                    <label htmlFor="kecamatan">Kecamatan</label>
-                    <input
-                      type="text"
-                      id="kecamatan"
-                      name="kecamatan"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                  <div>
-                    <label htmlFor="kota">Kabupaten/Kota</label>
-                    <input
-                      type="text"
-                      id="kota"
-                      name="kota"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                  <div>
-                    <label htmlFor="kodePos">Kode Pos</label>
-                    <input
-                      type="text"
-                      id="kodePos"
-                      name="kodePos"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="bg-white p-[20px] rounded-xl">
-            <p className="text-black font-regular text-lg mb-[10px]">
-              Silahkan lengkapi form berikut dengan data diri kepala cabang.
-            </p>
-            <div className="flex flex-col gap-2">
-              <div className="flex gap-3">
-                <div className="w-2/5">
-                  <label htmlFor="namaKepalaCabang">Nama Kepala Cabang</label>
-                  <input
-                    type="text"
-                    id="namaKepalaCabang"
-                    name="namaKepalaCabang"
-                    placeholder="Isi Nama Lengkap Kepala Cabang"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-                <div className="w-1/5 relative">
-                  <label htmlFor="statusPernikahan">Status Pernikahan</label>
-                  <select
-                    name="statusPernikahan"
-                    id="statusPernikahan"
-                    className="appearance-none w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-transparent focus:outline-none"
-                  >
-                    <option value="0">Pilih Status</option>
-                    <option value="1">Belum Menikah</option>
-                    <option value="2">Sudah Menikah</option>
-                  </select>
-                  <svg
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="absolute right-3 bottom-2.5 pointer-events-none"
-                  >
-                    <path
-                      d="M17.2902 9.31002C17.1977 9.21732 17.0878 9.14377 16.9668 9.09359C16.8459 9.04341 16.7162 9.01758 16.5852 9.01758C16.4543 9.01758 16.3246 9.04341 16.2036 9.09359C16.0826 9.14377 15.9727 9.21732 15.8802 9.31002L12.0002 13.19L8.12022 9.31002C7.93324 9.12304 7.67965 9.018 7.41522 9.018C7.1508 9.018 6.8972 9.12304 6.71022 9.31002C6.52324 9.497 6.4182 9.7506 6.4182 10.015C6.4182 10.2794 6.52324 10.533 6.71022 10.72L11.3002 15.31C11.3927 15.4027 11.5026 15.4763 11.6236 15.5265C11.7446 15.5766 11.8743 15.6025 12.0052 15.6025C12.1362 15.6025 12.2659 15.5766 12.3868 15.5265C12.5078 15.4763 12.6177 15.4027 12.7102 15.31L17.3002 10.72C17.6802 10.34 17.6802 9.70002 17.2902 9.31002Z"
-                      fill="black"
-                    />
-                  </svg>
-                </div>
-                <div className="w-2/5">
-                  <label htmlFor="namaSuamiIstri">Nama Suami/Istri</label>
-                  <input
-                    type="text"
-                    id="namaSuamiIstri"
-                    name="namaSuamiIstri"
-                    placeholder="*Jika Sudah Menikah"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-              </div>
-              <div className="flex gap-3">
-                <div className="w-1/4">
-                  <label htmlFor="tempatLahir">
-                    Isi Tempat Lahir Sesuai KTP
-                  </label>
-                  <input
-                    type="text"
-                    id="tempatLahir"
-                    name="tempatLahir"
-                    placeholder="*Jika Sudah Menikah"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-                <div className="w-1/4">
-                  <label htmlFor="tanggalLahir">Tanggal Lahir</label>
-                  <input
-                    type="date"
-                    id="tanggalLahir"
-                    name="tanggalLahir"
-                    placeholder="*Jika Sudah Menikah"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-                <div className="w-1/4">
-                  <label htmlFor="umur">Umur</label>
-                  <input
-                    type="text"
-                    id="umur"
-                    name="umur"
-                    placeholder="Auto generate"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    disabled
-                  />
-                </div>
-                <div className="w-1/5 relative">
-                  <label htmlFor="jenisKelamin">Jenis Kelamin</label>
-                  <select
-                    name="jenisKelamin"
-                    id="jenisKelamin"
-                    className="appearance-none w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  >
-                    <option value="1">Pilih Jenis Kelamin</option>
-                    <option value="1">Laki-laki</option>
-                    <option value="2">Perempuan</option>
-                  </select>
-                  <svg
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="absolute right-3 bottom-2.5 pointer-events-none"
-                  >
-                    <path
-                      d="M17.2902 9.31002C17.1977 9.21732 17.0878 9.14377 16.9668 9.09359C16.8459 9.04341 16.7162 9.01758 16.5852 9.01758C16.4543 9.01758 16.3246 9.04341 16.2036 9.09359C16.0826 9.14377 15.9727 9.21732 15.8802 9.31002L12.0002 13.19L8.12022 9.31002C7.93324 9.12304 7.67965 9.018 7.41522 9.018C7.1508 9.018 6.8972 9.12304 6.71022 9.31002C6.52324 9.497 6.4182 9.7506 6.4182 10.015C6.4182 10.2794 6.52324 10.533 6.71022 10.72L11.3002 15.31C11.3927 15.4027 11.5026 15.4763 11.6236 15.5265C11.7446 15.5766 11.8743 15.6025 12.0052 15.6025C12.1362 15.6025 12.2659 15.5766 12.3868 15.5265C12.5078 15.4763 12.6177 15.4027 12.7102 15.31L17.3002 10.72C17.6802 10.34 17.6802 9.70002 17.2902 9.31002Z"
-                      fill="black"
-                    />
-                  </svg>
-                </div>
-              </div>
-              <div className="flex gap-3">
-                <div className="w-2/5">
-                  <label htmlFor="nik">NIK</label>
-                  <input
-                    type="text"
-                    id="nik"
-                    name="nik"
-                    placeholder="NIK Harus Sesuai KTP dan Terdaftar di Dukcapil"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-                <div className="w-1/5 relative">
-                  <label htmlFor="agama">Agama</label>
-                  <select
-                    name="agama"
-                    id="agama"
-                    className="appearance-none w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  >
-                    <option value="0">Pilih Agama</option>
-                    <option value="1">Islam</option>
-                    <option value="2">Kristen</option>
-                    <option value="3">Katolik</option>
-                    <option value="4">Hindu</option>
-                    <option value="5">Buddha</option>
-                    <option value="6">Konghucu</option>
-                  </select>
-                  <svg
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="absolute right-3 bottom-2.5 pointer-events-none"
-                  >
-                    <path
-                      d="M17.2902 9.31002C17.1977 9.21732 17.0878 9.14377 16.9668 9.09359C16.8459 9.04341 16.7162 9.01758 16.5852 9.01758C16.4543 9.01758 16.3246 9.04341 16.2036 9.09359C16.0826 9.14377 15.9727 9.21732 15.8802 9.31002L12.0002 13.19L8.12022 9.31002C7.93324 9.12304 7.67965 9.018 7.41522 9.018C7.1508 9.018 6.8972 9.12304 6.71022 9.31002C6.52324 9.497 6.4182 9.7506 6.4182 10.015C6.4182 10.2794 6.52324 10.533 6.71022 10.72L11.3002 15.31C11.3927 15.4027 11.5026 15.4763 11.6236 15.5265C11.7446 15.5766 11.8743 15.6025 12.0052 15.6025C12.1362 15.6025 12.2659 15.5766 12.3868 15.5265C12.5078 15.4763 12.6177 15.4027 12.7102 15.31L17.3002 10.72C17.6802 10.34 17.6802 9.70002 17.2902 9.31002Z"
-                      fill="black"
-                    />
-                  </svg>
-                </div>
-                <div className="w-2/5">
-                  <label htmlFor="pekerjaan">Isi Sesuai KTP</label>
-                  <input
-                    type="text"
-                    id="pekerjaan"
-                    name="pekerjaan"
-                    placeholder="*Jika Sudah Menikah"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-              </div>
-              <div className="flex gap-3">
-                <div className="w-2/5">
-                  <div>
-                    <label htmlFor="alamatKepalaCabang">Alamat Lengkap</label>
-                    <input
-                      type="text"
-                      id="alamatCabang"
-                      name="alamatCabang"
-                      placeholder="Isi Alamat Lengkap Sesuai KTP"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                </div>
-                <div className="flex w-3/5 gap-3">
-                  <div>
-                    <label htmlFor="kelurahanKepalaCabang">
-                      Kelurahan/Desa
-                    </label>
-                    <input
-                      type="text"
-                      id="kelurahanKepalaCabang"
-                      name="kelurahanKepalaCabang"
-                      placeholder="Isi Sesuai KTP"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                  <div>
-                    <label htmlFor="kecamatanKepalaCabang">Kecamatan</label>
-                    <input
-                      type="text"
-                      id="kecamatanKepalaCabang"
-                      name="kecamatanKepalaCabang"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                  <div>
-                    <label htmlFor="kotaKepalaCabang">Kabupaten/Kota</label>
-                    <input
-                      type="text"
-                      id="kotaKepalaCabang"
-                      name="kotaKepalaCabang"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                  <div>
-                    <label htmlFor="kodePosKepalaCabang">Kode Pos</label>
-                    <input
-                      type="text"
-                      id="kodePosKepalaCabang"
-                      name="kodePosKepalaCabang"
-                      placeholder="Isi"
-                      className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                    />
-                  </div>
-                </div>
-              </div>
-              <div className="flex gap-3">
-                <div className="w-2/5 relative">
-                  <label htmlFor="agama">Pendidikan Terakhir</label>
-                  <select
-                    name="agama"
-                    id="agama"
-                    className="appearance-none w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  >
-                    <option value="0" className="">
-                      Pilih Pendidikan Terakhir
-                    </option>
-                    <option value="1">S1</option>
-                    <option value="2">SMA</option>
-                  </select>
-                  <svg
-                    width="24"
-                    height="24"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                    className="absolute right-3 bottom-2.5 pointer-events-none"
-                  >
-                    <path
-                      d="M17.2902 9.31002C17.1977 9.21732 17.0878 9.14377 16.9668 9.09359C16.8459 9.04341 16.7162 9.01758 16.5852 9.01758C16.4543 9.01758 16.3246 9.04341 16.2036 9.09359C16.0826 9.14377 15.9727 9.21732 15.8802 9.31002L12.0002 13.19L8.12022 9.31002C7.93324 9.12304 7.67965 9.018 7.41522 9.018C7.1508 9.018 6.8972 9.12304 6.71022 9.31002C6.52324 9.497 6.4182 9.7506 6.4182 10.015C6.4182 10.2794 6.52324 10.533 6.71022 10.72L11.3002 15.31C11.3927 15.4027 11.5026 15.4763 11.6236 15.5265C11.7446 15.5766 11.8743 15.6025 12.0052 15.6025C12.1362 15.6025 12.2659 15.5766 12.3868 15.5265C12.5078 15.4763 12.6177 15.4027 12.7102 15.31L17.3002 10.72C17.6802 10.34 17.6802 9.70002 17.2902 9.31002Z"
-                      fill="black"
-                    />
-                  </svg>
-                </div>
-                <div className="w-3/5">
-                  <label htmlFor="noHp">No. Hp (WhatsApp)</label>
-                  <input
-                    type="text"
-                    id="noHp"
-                    name="noHp"
-                    placeholder="Isi Nomor HP (WhatsApp)"
-                    className="w-full py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[10px] bg-white focus:outline-none"
-                  />
-                </div>
-              </div>
-              <div className="flex gap-3">
-                <div className="flex w-1/2">
-                  <Image
-                    src={"/images/image_none.jpg"}
-                    alt="Image_none"
-                    width={80}
-                    height={80}
-                    className="rounded-lg"
-                    style={{ objectFit: "cover" }}
-                  />
-                  <div className="flex flex-col ml-2 relative flex-grow">
-                    <label htmlFor="fotoDiri" className="text-start">
-                      Upload Foto Diri
-                    </label>
-                    <input
-                      type="file"
-                      name="fotoDiri"
-                      id="fotoDiri"
-                      className="py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[3px] bg-white focus:outline-none file:border-secondary file:border file:rounded-lg file:bg-transparent file:pr-[25px] file:py-[2px] file:mr-10 file:w-[180px] file:text-primary"
-                    />
-                    <svg
-                      width="17"
-                      height="16"
-                      viewBox="0 0 17 16"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="absolute left-[165px] bottom-5"
-                    >
-                      <path
-                        d="M3.25 14C2.78587 14 2.34075 13.8156 2.01256 13.4874C1.68437 13.1592 1.5 12.7141 1.5 12.25V9.75C1.5 9.55109 1.57902 9.36032 1.71967 9.21967C1.86032 9.07902 2.05109 9 2.25 9C2.44891 9 2.63968 9.07902 2.78033 9.21967C2.92098 9.36032 3 9.55109 3 9.75V12.25C3 12.388 3.112 12.5 3.25 12.5H13.75C13.8163 12.5 13.8799 12.4737 13.9268 12.4268C13.9737 12.3799 14 12.3163 14 12.25V9.75C14 9.55109 14.079 9.36032 14.2197 9.21967C14.3603 9.07902 14.5511 9 14.75 9C14.9489 9 15.1397 9.07902 15.2803 9.21967C15.421 9.36032 15.5 9.55109 15.5 9.75V12.25C15.5 12.7141 15.3156 13.1592 14.9874 13.4874C14.6592 13.8156 14.2141 14 13.75 14H3.25Z"
-                        fill="#004080"
-                      />
-                      <path
-                        d="M12.2795 4.72C12.3491 4.7896 12.4043 4.87223 12.442 4.96316C12.4797 5.0541 12.4991 5.15157 12.4991 5.25C12.4991 5.34843 12.4797 5.44589 12.442 5.53683C12.4043 5.62777 12.3491 5.7104 12.2795 5.78C12.2099 5.8496 12.1273 5.90481 12.0364 5.94248C11.9454 5.98014 11.848 5.99953 11.7495 5.99953C11.6511 5.99953 11.5536 5.98014 11.4627 5.94248C11.3718 5.90481 11.2891 5.8496 11.2195 5.78L9.24953 3.811V9.5C9.24953 9.69891 9.17052 9.88968 9.02986 10.0303C8.88921 10.171 8.69845 10.25 8.49953 10.25C8.30062 10.25 8.10986 10.171 7.9692 10.0303C7.82855 9.88968 7.74953 9.69891 7.74953 9.5V3.811L5.77953 5.78C5.70993 5.8496 5.6273 5.90481 5.53637 5.94248C5.44543 5.98014 5.34796 5.99953 5.24953 5.99953C5.1511 5.99953 5.05364 5.98014 4.9627 5.94248C4.87176 5.90481 4.78913 5.8496 4.71953 5.78C4.64993 5.7104 4.59472 5.62777 4.55705 5.53683C4.51939 5.44589 4.5 5.34843 4.5 5.25C4.5 5.15157 4.51939 5.0541 4.55705 4.96316C4.59472 4.87223 4.64993 4.7896 4.71953 4.72L7.96953 1.47C8.0391 1.40033 8.12171 1.34507 8.21266 1.30736C8.3036 1.26965 8.40108 1.25024 8.49953 1.25024C8.59798 1.25024 8.69547 1.26965 8.78641 1.30736C8.87735 1.34507 8.95997 1.40033 9.02953 1.47L12.2795 4.72Z"
-                        fill="#004080"
-                      />
-                    </svg>
-                  </div>
-                </div>
-                <div className="flex w-1/2">
-                  <Image
-                    src={"/images/image_none.jpg"}
-                    alt="Image_none"
-                    width={80}
-                    height={80}
-                    style={{ objectFit: "cover" }}
-                    className="rounded-lg"
-                  />
-                  <div className="flex flex-col ml-2 relative flex-grow">
-                    <label htmlFor="fotoKtp" className="text-start">
-                      Upload Foto KTP
-                    </label>
-                    <input
-                      type="file"
-                      name="fotoKtp"
-                      id="fotoKtp"
-                      className="py-[10px] px-[20px] border border-[#d9d9d9] rounded-md mt-[3px] bg-white focus:outline-none file:border-secondary file:border file:rounded-lg file:bg-transparent file:pr-[25px] file:py-[2px] file:mr-10 file:w-[180px] file:text-primary"
-                    />
-                    <svg
-                      width="17"
-                      height="16"
-                      viewBox="0 0 17 16"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                      className="absolute left-[165px] bottom-5"
-                    >
-                      <path
-                        d="M3.25 14C2.78587 14 2.34075 13.8156 2.01256 13.4874C1.68437 13.1592 1.5 12.7141 1.5 12.25V9.75C1.5 9.55109 1.57902 9.36032 1.71967 9.21967C1.86032 9.07902 2.05109 9 2.25 9C2.44891 9 2.63968 9.07902 2.78033 9.21967C2.92098 9.36032 3 9.55109 3 9.75V12.25C3 12.388 3.112 12.5 3.25 12.5H13.75C13.8163 12.5 13.8799 12.4737 13.9268 12.4268C13.9737 12.3799 14 12.3163 14 12.25V9.75C14 9.55109 14.079 9.36032 14.2197 9.21967C14.3603 9.07902 14.5511 9 14.75 9C14.9489 9 15.1397 9.07902 15.2803 9.21967C15.421 9.36032 15.5 9.55109 15.5 9.75V12.25C15.5 12.7141 15.3156 13.1592 14.9874 13.4874C14.6592 13.8156 14.2141 14 13.75 14H3.25Z"
-                        fill="#004080"
-                      />
-                      <path
-                        d="M12.2795 4.72C12.3491 4.7896 12.4043 4.87223 12.442 4.96316C12.4797 5.0541 12.4991 5.15157 12.4991 5.25C12.4991 5.34843 12.4797 5.44589 12.442 5.53683C12.4043 5.62777 12.3491 5.7104 12.2795 5.78C12.2099 5.8496 12.1273 5.90481 12.0364 5.94248C11.9454 5.98014 11.848 5.99953 11.7495 5.99953C11.6511 5.99953 11.5536 5.98014 11.4627 5.94248C11.3718 5.90481 11.2891 5.8496 11.2195 5.78L9.24953 3.811V9.5C9.24953 9.69891 9.17052 9.88968 9.02986 10.0303C8.88921 10.171 8.69845 10.25 8.49953 10.25C8.30062 10.25 8.10986 10.171 7.9692 10.0303C7.82855 9.88968 7.74953 9.69891 7.74953 9.5V3.811L5.77953 5.78C5.70993 5.8496 5.6273 5.90481 5.53637 5.94248C5.44543 5.98014 5.34796 5.99953 5.24953 5.99953C5.1511 5.99953 5.05364 5.98014 4.9627 5.94248C4.87176 5.90481 4.78913 5.8496 4.71953 5.78C4.64993 5.7104 4.59472 5.62777 4.55705 5.53683C4.51939 5.44589 4.5 5.34843 4.5 5.25C4.5 5.15157 4.51939 5.0541 4.55705 4.96316C4.59472 4.87223 4.64993 4.7896 4.71953 4.72L7.96953 1.47C8.0391 1.40033 8.12171 1.34507 8.21266 1.30736C8.3036 1.26965 8.40108 1.25024 8.49953 1.25024C8.59798 1.25024 8.69547 1.26965 8.78641 1.30736C8.87735 1.34507 8.95997 1.40033 9.02953 1.47L12.2795 4.72Z"
-                        fill="#004080"
-                      />
-                    </svg>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-
-          <button
-            type="submit"
-            className="w-[148px] px-[20px] py-[12px] text-white bg-primary rounded-lg ml-auto"
-          >
-            Simpan
-          </button>
-        </div>
-      </form>
-    </main>
-  );
-};
-
-export default BuatAkunCabang;
