+++ app/data-master/layout.jsx
@@ -5,9 +5,7 @@ const Layout = ({ children }) => {
   return (
     <main className="flex h-full">
       <Sidebar />
-      <div className="m-[30px] flex flex-col gap-[20px] w-full ml-[310px]">
-        {children}
-      </div>
+      <div className="m-[30px] w-full ml-[310px]">{children}</div>
     </main>
   );
 };
